feat(paramparser): Add Declaration Objects
Added objects for declarative statements.
Finished RapStatementFactory

+++ BisUtils.Parsers/BisUtils.Parsers.ParamParser/Factories/RapStatementFactory.cs
@@ -1,3 +1,4 @@
+using BisUtils.Parsers.ParamParser.Declarations;
 using BisUtils.Parsers.ParamParser.Interfaces;
 using BisUtils.Parsers.ParamParser.Statements;
 
@@ -5,8 +6,11 @@ namespace BisUtils.Parsers.ParamParser.Factories;
 
 public class RapStatementFactory {
     public static IRapStatement Create(Generated.ParamLang.ParamParser.StatementContext ctx) {
+        if (ctx.classDeclaration() is { } @class) return (IRapStatement) RapClassDeclaration.FromContext(@class);
         if (ctx.externalClassDeclaration() is { } external) return RapExternalClassStatement.FromContext(external);
+        if (ctx.tokenDeclaration() is { } var) return RapVariableDeclaration.FromContext(var);
         if (ctx.arrayAppension() is { } appension) return RapAppensionStatement.FromContext(appension);
+        if (ctx.arrayDeclaration() is { } array) return  RapArrayDeclaration.FromContext(array);
         if (ctx.deleteStatement() is { } delete) return RapDeleteStatement.FromContext(delete);
         throw new NotSupportedException();
     }
